[#15] album_detail_screen 코드 분리
