+++ lib/presentation/screens/main/detail/album_detail_screen.dart
@@ -6,6 +6,8 @@ import 'package:provider/provider.dart';
 import 'album_detail_view_model.dart';
 import 'package:cherrypic/core/constants/font.dart';
 import 'package:cherrypic/core/constants/color.dart';
+part 'parts/album_detail_segmented.dart';
+part 'parts/album_detail_selecting_bar.dart';
 
 class AlbumDetailScreen extends StatelessWidget {
   final int albumId;
@@ -145,220 +147,6 @@ class _BodyState extends State<_Body> {
                   );
                 },
               ),
-            ),
-          ),
-        ],
-      ),
-    );
-  }
-}
-
-/// 하단 토글: 전체/이벤트
-class _FloatingSegmented extends StatelessWidget {
-  const _FloatingSegmented({
-    required this.value, // 0=전체, 1=이벤트
-    required this.onChanged,
-  });
-
-  final int value;
-  final ValueChanged<int> onChanged;
-
-  @override
-  Widget build(BuildContext context) {
-    // 트랙/핸들 규격
-    const double trackW = 120;
-    const double trackH = 50;
-    const double handleH = 35;
-    const double edge = 5;
-    const double leftHandleW = 50; // 전체
-    const double rightHandleW = 60; // 이벤트
-
-    final double handleW = value == 0 ? leftHandleW : rightHandleW;
-    final double handleLeft = value == 0 ? edge : trackW - edge - handleW;
-
-    // 각 절반의 중앙 좌표(라벨 기본 위치)
-    const double centerLeft = trackW / 4; // 30
-    const double centerRight = trackW * 3 / 4; // 90
-
-    // 선택된 핸들의 실제 중앙 좌표
-    final double handleCenter = value == 0
-        ? edge + leftHandleW / 2
-        : trackW - edge - rightHandleW / 2;
-
-    // 라벨 보정: 선택된 쪽만 핸들 중심에 맞춰 살짝 이동
-    final double dxLeftLabel = value == 0
-        ? (handleCenter - centerLeft)
-        : 0.0; // 0
-    final double dxRightLabel = value == 1
-        ? (handleCenter - centerRight)
-        : 0.0; // -5
-
-    return SizedBox(
-      width: trackW,
-      height: trackH,
-      child: Stack(
-        children: [
-          // 트랙
-          Container(
-            width: trackW,
-            height: trackH,
-            decoration: BoxDecoration(
-              color: AppColor.mainLightRed.withAlpha(70),
-              borderRadius: BorderRadius.circular(25),
-            ),
-          ),
-
-          // 선택 핸들 (폭/위치 애니메이션)
-          AnimatedPositioned(
-            duration: const Duration(milliseconds: 220),
-            curve: Curves.easeOut,
-            left: handleLeft,
-            top: (trackH - handleH) / 2,
-            width: handleW,
-            height: handleH,
-            child: DecoratedBox(
-              decoration: BoxDecoration(
-                color: AppColor.mainRed,
-                borderRadius: BorderRadius.circular(20),
-              ),
-            ),
-          ),
-
-          // 라벨 + 터치
-          Row(
-            children: [
-              Expanded(
-                child: InkWell(
-                  borderRadius: BorderRadius.circular(25),
-                  onTap: () => onChanged(0),
-                  child: Center(
-                    child: Transform.translate(
-                      offset: Offset(dxLeftLabel, 0),
-                      child: Text(
-                        '전체',
-                        style: AppFont.size16.copyWith(
-                          fontWeight: FontWeight.w700,
-                          color: value == 0 ? Colors.white : AppColor.mainRed,
-                        ),
-                      ),
-                    ),
-                  ),
-                ),
-              ),
-              Expanded(
-                child: InkWell(
-                  borderRadius: BorderRadius.circular(25),
-                  onTap: () => onChanged(1),
-                  child: Center(
-                    child: Transform.translate(
-                      offset: Offset(dxRightLabel, 0),
-                      child: Text(
-                        '이벤트',
-                        style: AppFont.size16.copyWith(
-                          fontWeight: FontWeight.w700,
-                          color: value == 1 ? Colors.white : AppColor.mainRed,
-                        ),
-                      ),
-                    ),
-                  ),
-                ),
-              ),
-            ],
-          ),
-        ],
-      ),
-    );
-  }
-}
-
-/// 하단 우측 사진 추가 버튼
-class _AddPhotoButton extends StatelessWidget {
-  const _AddPhotoButton({required this.onTap});
-  final VoidCallback onTap;
-
-  @override
-  Widget build(BuildContext context) {
-    return GestureDetector(
-      onTap: onTap,
-      child: Container(
-        width: 50,
-        height: 50,
-        alignment: Alignment.center,
-        decoration: BoxDecoration(
-          color: AppColor.mainRed,
-          shape: BoxShape.circle,
-          boxShadow: [
-            BoxShadow(
-              color: Colors.black.withAlpha(15),
-              blurRadius: 12,
-              offset: const Offset(0, 6),
-            ),
-          ],
-        ),
-        child: Image.asset(
-          'assets/images/img_add_bt.png',
-          width: 25,
-          height: 25,
-          fit: BoxFit.contain,
-        ),
-      ),
-    );
-  }
-}
-
-class _SelectingBar extends StatelessWidget {
-  final int count;
-  final VoidCallback onMore;
-  const _SelectingBar({super.key, required this.count, required this.onMore});
-
-  @override
-  Widget build(BuildContext context) {
-    return Container(
-      height: 40,
-      padding: const EdgeInsets.symmetric(horizontal: 12),
-      decoration: BoxDecoration(
-        color: AppColor.mainRed.withAlpha(150),
-        borderRadius: BorderRadius.circular(20),
-        boxShadow: [
-          BoxShadow(
-            color: Colors.black.withAlpha(38),
-            blurRadius: 12,
-            offset: const Offset(0, 6),
-          ),
-        ],
-      ),
-      child: Row(
-        mainAxisSize: MainAxisSize.min,
-        children: [
-          // "#장 선택됨" 캡슐(흰색 1px 테두리)
-          Container(
-            height: 28,
-            alignment: Alignment.center,
-            padding: const EdgeInsets.symmetric(horizontal: 12),
-            decoration: BoxDecoration(
-              borderRadius: BorderRadius.circular(16),
-              border: Border.all(color: Colors.white, width: 1),
-            ),
-            child: Text(
-              '$count 장 선택됨',
-              style: AppFont.size18.copyWith(
-                color: Colors.white,
-                fontWeight: FontWeight.w700,
-              ),
-            ),
-          ),
-          const SizedBox(width: 8),
-          // 더보기 버튼
-          GestureDetector(
-            onTap: onMore,
-            child: Container(
-              width: 28,
-              height: 28,
-              decoration: const BoxDecoration(
-                color: AppColor.mainLightRed,
-                shape: BoxShape.circle,
-              ),
-              child: Icon(Icons.more_horiz, size: 18, color: AppColor.mainRed),
             ),
           ),
         ],
